chore: removed unnecessary split in files
